backend/requirements/requirements.txt
fastapi>=0.115
uvicorn>=0.22
pydantic>=2
python-socketio>=5.13
python-engineio>=4.12
simple-websocket>=1.0
psutil>=5.9
gputil>=1.4.0
pymilvus>=2.4.0
pymongo>=4.7.0


# Training Stuff
torch==2.7.1+cu118
torchvision==0.22.1+cu118
torchaudio==2.7.1+cu118
--extra-index-url https://download.pytorch.org/whl/cu118

filterpy
huggingface-hub

pycuda==2024.1
transformers>=4.45.2
accelerate>=1.0.0
pillow>=10.3.0
safetensors
opencv-python==4.10.0.84
numpy==1.26.4
tqdm>=4.66.4
einops
timm

onnx>=1.15,<2.0
onnxruntime-gpu==1.16.3

scikit-image
matplotlib